requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
openai>=0.27.0
pydantic>=1.8.0
python-dotenv>=0.19.0
pandas>=1.3.0
prophet>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dateutil>=2.8.2
plotly>=5.3.1
langchain-openai
langchain-community
langchain-core
